refactor: fee_rate statistics
