+++ rpc/src/util/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod fee_rate;
+
+pub(crate) use fee_rate::FeeRateCollector;
